requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
gunicorn
whitenoise
